handling context timeout/cancelled in printer package
